fix(web): show friendly connector type instead of raw kind
Both the list and manage pages showed the raw kind ("GOOGLE",
"MICROSOFT") as the connector's type badge, and the manage page
matched a preset by kind alone — picking the first preset of that
kind regardless of scopes, so e.g. a Google Docs connector showed as
Gmail. Move matchesPreset into connectorPresets.ts as presetFor, used
by both pages, and render preset.name (uppercased) instead of kind.
